Revert "Refactor"
This reverts commit 5787076cb3e8166c6e220adc4fcfe7d3dd637cc6.
