+++ SplitWiseImpl/src/com/dataStore/UserDataStore.java
@@ -2,8 +2,7 @@ package com.dataStore;
 
 import java.util.HashMap;
 
-import com.dataStore.interfaces.IUserDataStore;
-import com.user.UserDetails;
+import com.interfaces.IUserDataStore;
 
 public class UserDataStore implements IUserDataStore {
 	HashMap<String, UserDetails> userDetailMap = new HashMap<String, UserDetails>();
