--- SplitWiseImpl/src/com/factory/interfaces/IDataStoreFactory.java
@@ -1,10 +1,0 @@
-package com.factory.interfaces;
-
-import com.dataStore.interfaces.IUserDataStore;
-import com.dataStore.interfaces.IUserPairOwedAmountDataStore;
-
-public interface IDataStoreFactory {
-	IUserDataStore getUserDataStore();
-	
-	IUserPairOwedAmountDataStore getUserPairOwedAmountDataStore();
-}
