+++ SplitWiseImpl/src/com/interfaces/IDataStoreFactory.java
@@ -1,0 +1,7 @@
+package com.interfaces;
+
+public interface IDataStoreFactory {
+	IUserDataStore getUserDataStore();
+	
+	IUserPairOwedAmountDataStore getUserPairOwedAmountDataStore();
+}
