+++ SplitWiseImpl/src/com/interfaces/IStrategyListFactory.java
@@ -1,8 +1,6 @@
-package com.factory.interfaces;
+package com.interfaces;
 
 import java.util.List;
-
-import com.interfaces.IStrategyList;
 
 public interface IStrategyListFactory {
 	IStrategyList getStrategyList(List<String> strategiesPresent);
