--- SplitWiseImpl/src/com/ioParser/ResultFormatter.java
@@ -1,5 +1,0 @@
-package com.ioParser;
-
-public class ResultFormatter {
-
-}
