+++ SplitWiseImpl/src/com/percentDivisionStrategy/PercentDivisionInputGenerator.java
@@ -1,11 +1,10 @@
-package com.strategies.percentDivisionStrategy;
+package com.percentDivisionStrategy;
 
 import java.util.HashMap;
 import java.util.List;
-
-import com.strategies.interfaces.ISplitInput;
-import com.strategies.splitObjects.SplitInput;
-import com.strategies.splitObjects.StrategyInputGenerator;
+import com.interfaces.ISplitInput;
+import com.splitObjects.SplitInput;
+import com.splitObjects.StrategyInputGenerator;
 
 public class PercentDivisionInputGenerator extends StrategyInputGenerator {
 	@Override
