+++ SplitWiseImpl/src/com/splitObjects/SplitResult.java
@@ -1,4 +1,4 @@
-package com.strategies.splitObjects;
+package com.splitObjects;
 
 import java.util.HashMap;
